move `SignatureTypeParameterProcessor` to separate file
so we can reuse it to process method type parameters.
